+++ packages/liveblocks-react/src/__tests__/index.test.tsx
@@ -185,6 +185,7 @@ describe("useRoom", () => {
       JSON.stringify([
         {
           type: ClientMsgCode.UPDATE_PRESENCE,
+          targetActor: -1,
           data: { x: 1 },
         },
       ])